Fix login unit test when input valid form data #7
