+++ .circleci/config.yml
@@ -40,6 +40,7 @@ jobs:
           name: run tests
           command: |
             bundle exec rake test 
+            npm test
             bundle exec rails s -d 
             kill `cat tmp/pids/server.pid`
