+++ .gitignore
@@ -4,6 +4,9 @@
 # or operating system, you probably want to add a global ignore instead:
 #   git config --global core.excludesfile '~/.gitignore_global'
 
+# Ignore compile js
+/app/assets/javascripts/applicaiton.js
+
 # Ignore bundler config.
 /.bundle
