+++ frontend/test/jest/test-repo/index.html
@@ -20,28 +20,28 @@
     </h1>
     <div class='clearfix'>
       <div class='fl pad1y space-right2'>
-        <span class="strong">44.62% </span>
+        <span class="strong">51.14% </span>
         <span class="quiet">Statements</span>
-        <span class='fraction'>87/195</span>
+        <span class='fraction'>112/219</span>
       </div>
       <div class='fl pad1y space-right2'>
-        <span class="strong">5.56% </span>
+        <span class="strong">11.11% </span>
         <span class="quiet">Branches</span>
-        <span class='fraction'>1/18</span>
+        <span class='fraction'>2/18</span>
       </div>
       <div class='fl pad1y space-right2'>
-        <span class="strong">54.17% </span>
+        <span class="strong">61.54% </span>
         <span class="quiet">Functions</span>
-        <span class='fraction'>13/24</span>
+        <span class='fraction'>16/26</span>
       </div>
       <div class='fl pad1y space-right2'>
-        <span class="strong">44.85% </span>
+        <span class="strong">51.38% </span>
         <span class="quiet">Lines</span>
-        <span class='fraction'>87/194</span>
+        <span class='fraction'>112/218</span>
       </div>
     </div>
   </div>
-  <div class='status-line low'></div>
+  <div class='status-line medium'></div>
 <div class="pad1">
 <table class="coverage-summary">
 <thead>
@@ -60,15 +60,15 @@
 </thead>
 <tbody><tr>
 	<td class="file medium" data-value="root"><a href="root/index.html">root</a></td>
-	<td data-value="50" class="pic medium"><div class="chart"><div class="cover-fill" style="width: 50%;"></div><div class="cover-empty" style="width:50%;"></div></div></td>
+	<td data-value="58.33" class="pic medium"><div class="chart"><div class="cover-fill" style="width: 58%;"></div><div class="cover-empty" style="width:42%;"></div></div></td>
+	<td data-value="58.33" class="pct medium">58.33%</td>
+	<td data-value="12" class="abs medium">7/12</td>
 	<td data-value="50" class="pct medium">50%</td>
-	<td data-value="12" class="abs medium">6/12</td>
-	<td data-value="25" class="pct low">25%</td>
-	<td data-value="4" class="abs low">1/4</td>
-	<td data-value="0" class="pct low">0%</td>
-	<td data-value="1" class="abs low">0/1</td>
-	<td data-value="50" class="pct medium">50%</td>
-	<td data-value="12" class="abs medium">6/12</td>
+	<td data-value="4" class="abs medium">2/4</td>
+	<td data-value="100" class="pct high">100%</td>
+	<td data-value="1" class="abs high">1/1</td>
+	<td data-value="58.33" class="pct medium">58.33%</td>
+	<td data-value="12" class="abs medium">7/12</td>
 	</tr>
 
 <tr>
@@ -88,13 +88,13 @@
 	<td class="file high" data-value="root/Users/kensei/program/ruby/todo-app-v2/frontend/test/jest/Login"><a href="root/Users/kensei/program/ruby/todo-app-v2/frontend/test/jest/Login/index.html">root/Users/kensei/program/ruby/todo-app-v2/frontend/test/jest/Login</a></td>
 	<td data-value="100" class="pic high"><div class="chart"><div class="cover-fill cover-full" style="width: 100%;"></div><div class="cover-empty" style="width:0%;"></div></div></td>
 	<td data-value="100" class="pct high">100%</td>
-	<td data-value="31" class="abs high">31/31</td>
+	<td data-value="55" class="abs high">55/55</td>
 	<td data-value="100" class="pct high">100%</td>
 	<td data-value="0" class="abs high">0/0</td>
 	<td data-value="100" class="pct high">100%</td>
-	<td data-value="4" class="abs high">4/4</td>
+	<td data-value="6" class="abs high">6/6</td>
 	<td data-value="100" class="pct high">100%</td>
-	<td data-value="31" class="abs high">31/31</td>
+	<td data-value="55" class="abs high">55/55</td>
 	</tr>
 
 <tr>
@@ -116,7 +116,7 @@
 </div><!-- /wrapper -->
 <div class='footer quiet pad2 space-top1 center small'>
   Code coverage
-  generated by <a href="https://istanbul.js.org/" target="_blank">istanbul</a> at Mon Jan 22 2018 17:37:02 GMT+0900 (JST)
+  generated by <a href="https://istanbul.js.org/" target="_blank">istanbul</a> at Tue Jan 23 2018 18:39:22 GMT+0900 (JST)
 </div>
 </div>
 <script src="prettify.js"></script>
