--- frontend/test/jest/test-repo/root/Users/kensei/program/ruby/todo-app-v2/frontend/src/javascripts/store/index.html
@@ -1,93 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <title>Code coverage report for root/Users/kensei/program/ruby/todo-app-v2/frontend/src/javascripts/store</title>
-    <meta charset="utf-8" />
-    <link rel="stylesheet" href="../../../../../../../../../../prettify.css" />
-    <link rel="stylesheet" href="../../../../../../../../../../base.css" />
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <style type='text/css'>
-        .coverage-summary .sorter {
-            background-image: url(../../../../../../../../../../sort-arrow-sprite.png);
-        }
-    </style>
-</head>
-<body>
-<div class='wrapper'>
-  <div class='pad1'>
-    <h1>
-      <a href="../../../../../../../../../../index.html">All files</a> root/Users/kensei/program/ruby/todo-app-v2/frontend/src/javascripts/store
-    </h1>
-    <div class='clearfix'>
-      <div class='fl pad1y space-right2'>
-        <span class="strong">3.85% </span>
-        <span class="quiet">Statements</span>
-        <span class='fraction'>4/104</span>
-      </div>
-      <div class='fl pad1y space-right2'>
-        <span class="strong">0% </span>
-        <span class="quiet">Branches</span>
-        <span class='fraction'>0/14</span>
-      </div>
-      <div class='fl pad1y space-right2'>
-        <span class="strong">25% </span>
-        <span class="quiet">Functions</span>
-        <span class='fraction'>3/12</span>
-      </div>
-      <div class='fl pad1y space-right2'>
-        <span class="strong">3.88% </span>
-        <span class="quiet">Lines</span>
-        <span class='fraction'>4/103</span>
-      </div>
-    </div>
-  </div>
-  <div class='status-line low'></div>
-<div class="pad1">
-<table class="coverage-summary">
-<thead>
-<tr>
-   <th data-col="file" data-fmt="html" data-html="true" class="file">File</th>
-   <th data-col="pic" data-type="number" data-fmt="html" data-html="true" class="pic"></th>
-   <th data-col="statements" data-type="number" data-fmt="pct" class="pct">Statements</th>
-   <th data-col="statements_raw" data-type="number" data-fmt="html" class="abs"></th>
-   <th data-col="branches" data-type="number" data-fmt="pct" class="pct">Branches</th>
-   <th data-col="branches_raw" data-type="number" data-fmt="html" class="abs"></th>
-   <th data-col="functions" data-type="number" data-fmt="pct" class="pct">Functions</th>
-   <th data-col="functions_raw" data-type="number" data-fmt="html" class="abs"></th>
-   <th data-col="lines" data-type="number" data-fmt="pct" class="pct">Lines</th>
-   <th data-col="lines_raw" data-type="number" data-fmt="html" class="abs"></th>
-</tr>
-</thead>
-<tbody><tr>
-	<td class="file low" data-value="store.js"><a href="store.js.html">store.js</a></td>
-	<td data-value="3.85" class="pic low"><div class="chart"><div class="cover-fill" style="width: 3%;"></div><div class="cover-empty" style="width:97%;"></div></div></td>
-	<td data-value="3.85" class="pct low">3.85%</td>
-	<td data-value="104" class="abs low">4/104</td>
-	<td data-value="0" class="pct low">0%</td>
-	<td data-value="14" class="abs low">0/14</td>
-	<td data-value="25" class="pct low">25%</td>
-	<td data-value="12" class="abs low">3/12</td>
-	<td data-value="3.88" class="pct low">3.88%</td>
-	<td data-value="103" class="abs low">4/103</td>
-	</tr>
-
-</tbody>
-</table>
-</div><div class='push'></div><!-- for sticky footer -->
-</div><!-- /wrapper -->
-<div class='footer quiet pad2 space-top1 center small'>
-  Code coverage
-  generated by <a href="https://istanbul.js.org/" target="_blank">istanbul</a> at Wed Jan 17 2018 10:39:10 GMT+0900 (JST)
-</div>
-</div>
-<script src="../../../../../../../../../../prettify.js"></script>
-<script>
-window.onload = function () {
-        if (typeof prettyPrint === 'function') {
-            prettyPrint();
-        }
-};
-</script>
-<script src="../../../../../../../../../../sorter.js"></script>
-</body>
-</html>
