+++ frontend/test/jest/test-repo/root/Users/kensei/program/ruby/todo-app-v2/frontend/test/jest/App/App.spec.js.html
@@ -127,7 +127,9 @@
 82
 83
 84
-85</td><td class="line-coverage quiet"><span class="cline-any cline-neutral">&nbsp;</span>
+85
+86</td><td class="line-coverage quiet"><span class="cline-any cline-neutral">&nbsp;</span>
+<span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
@@ -215,6 +217,7 @@
 import Vuex from 'vuex'
 import App from '../../../src/javascripts/App.vue'
 import module from '../store/store_test'
+import { createRenderer } from 'vue-server-renderer'
 const localVue = createLocalVue()
 &nbsp;
 localVue.use(Vuex)
@@ -243,7 +246,7 @@ describe('App', () =&gt; {
     store = new Vuex.Store({
       state: {},
       actions,
-      getters
+      getters: getters
     })
   })
   
@@ -301,7 +304,7 @@ describe('App', () =&gt; {
 </div><!-- /wrapper -->
 <div class='footer quiet pad2 space-top1 center small'>
   Code coverage
-  generated by <a href="https://istanbul.js.org/" target="_blank">istanbul</a> at Mon Jan 22 2018 17:37:02 GMT+0900 (JST)
+  generated by <a href="https://istanbul.js.org/" target="_blank">istanbul</a> at Tue Jan 23 2018 18:39:22 GMT+0900 (JST)
 </div>
 </div>
 <script src="../../../../../../../../../../prettify.js"></script>
