+++ frontend/test/jest/test-repo/root/Users/kensei/program/ruby/todo-app-v2/frontend/test/jest/store/index.html
@@ -77,7 +77,7 @@
 </div><!-- /wrapper -->
 <div class='footer quiet pad2 space-top1 center small'>
   Code coverage
-  generated by <a href="https://istanbul.js.org/" target="_blank">istanbul</a> at Mon Jan 22 2018 17:37:02 GMT+0900 (JST)
+  generated by <a href="https://istanbul.js.org/" target="_blank">istanbul</a> at Tue Jan 23 2018 18:39:22 GMT+0900 (JST)
 </div>
 </div>
 <script src="../../../../../../../../../../prettify.js"></script>
